+++ src/DVD.java
@@ -7,7 +7,7 @@ public class DVD extends Produit {
     public DVD(){
         super();
     }
-    public DVD(int prix, String produit, int pointsFidelite){
+    public DVD(double prix, String produit, int pointsFidelite){
         super(prix, produit, pointsFidelite);
     }
 }
