+++ src/Produit.java
@@ -6,21 +6,21 @@ import java.util.ArrayList;
 public abstract class Produit {
 
     protected String titre;
-    protected float prix;
+    protected double prix;
     protected ArrayList<Offre> sesOffres;
     protected int pointsFidelite;
 
     public Produit(){
         this(-1, "", 0);
     }
-    public Produit(float prix, String titre, int pointsFidelite){
+    public Produit(double prix, String titre, int pointsFidelite){
         this.prix = prix;
         this.titre = titre;
         this.pointsFidelite = pointsFidelite;
         sesOffres = new ArrayList<>();
     }
 
-    public float getPrix() {
+    public double getPrix() {
         return prix;
     }
     public String getTitre(){
